feat(folder-view): implement validation
Remove unused libraries

Add validation to the get_object function. Now it returns the folder if the user and id_folder match in the db.

Add validation when deleting a folder. Now it verifies that the folder is not the default, if it is, it will raise an error.

+++ passwordManagerAPI/views/folders.py
@@ -1,12 +1,7 @@
 from passwordManagerAPI.models import Folder, Items
-from django.contrib.auth.models import User
 from passwordManagerAPI.serializers import FolderSerializer
-from rest_framework import viewsets, status
+from rest_framework import status
 from rest_framework.views import APIView
-from rest_framework.authentication import TokenAuthentication
-from rest_framework.permissions import IsAuthenticated
-from django.contrib.auth.decorators import login_required
-from rest_framework.authtoken.views import Token
 from django.http import Http404
 from rest_framework.response import Response
 
@@ -34,10 +29,12 @@ class FolderView(APIView):
 
 class DetailFolderView(APIView):
     def get_object(self, pk):
-        try:
-            return Folder.objects.get(pk=pk)
-        except Folder.DoesNotExist:
+        folder = Folder.objects.filter(
+            user=self.request.user.id, pk=pk).first()
+
+        if folder is None:
             raise Http404
+        return folder
 
     def get(self, request, pk, format=None):
         folder = self.get_object(pk)
@@ -59,8 +56,14 @@ class DetailFolderView(APIView):
 
     def delete(self, request, pk, format=None):
         folder = self.get_object(pk)
-        default = Folder.objects.filter(user=self.request.user).first()
-        item = Items.objects.filter(folder=folder).update(folder=default)
+
+        if folder.verify_default_folder():
+            return Response(status=status.HTTP_400_BAD_REQUEST, message='You cannot delete the default folder.')
+
+        default = Folder.objects.filter(
+            user=self.request.user, name='default').first()
+        item = Items.objects.filter(
+            folder=folder, user=self.request.user).update(folder=default)
         folder.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
